refactor: organize numeric readers
