+++ src/lexer/number/BinaryReader.js
