+++ src/lexer/number/HexReader.js
